+++ CHANGELOG.md
@@ -22,7 +22,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
-- Nothing yet.
+- [GHSA-w476-p2h3-79g9](https://github.com/advisories/GHSA-w476-p2h3-79g9) determined to
+  not be a vulnerability in *aquarion-libtts* itself.  Only the development environment
+  is impacted, and even there, the risk is low.  This will be fixed in the near future.
 
 ### Added
 
@@ -59,7 +61,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
-- Nothing, first release.
+- [CVE-2025-8869](https://nvd.nist.gov/vuln/detail/cve-2025-8869) determined to not be
+  a risk because the minimum version Python version of *aquarion-libtts* is 3.12.
 
 ### Added