+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support for Python 3.13.
+- A Software Bill of Materials (SBOM) in CycloneDX format.
 
 ### Fixed