+++ CHANGELOG.md
@@ -18,11 +18,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.1.1 - 2025-10-23
+## 0.1.1 - 2025-11-05
 
 ### Fixed
 
 - Fixed `en-core-web-sm` dependency failure when trying to pip install aquarion-libtts.
+  This is a first pass, quick fix.  A more comprehensive solution will be included in a
+  future version.
 - Fixed `TTSPluginRegistry.load_plugins` not working without the `kokoro` extra.
 
 ----